% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{accessor}
\alias{accessor}
\alias{data_dim}
\alias{method}
\alias{n_obs}
\alias{n_obs_tot}
\alias{false_alarm_probability}
\alias{leading_constant}
\alias{baseline_sd}
\alias{statistics}
\alias{all_statistics}
\alias{status}
\alias{bool_estimate_mean}
\title{Accessor functions for the 'CHAD' class}
\usage{
data_dim(detector)

method(detector)

n_obs(detector)

n_obs_tot(detector)

false_alarm_probability(detector)

leading_constant(detector)

baseline_sd(detector)

statistics(detector)

all_statistics(detector)

status(detector)

bool_estimate_mean(detector)
}
\arguments{
\item{detector}{Object of class 'CHAD'}
}
\description{
Accessor functions for the 'CHAD' class
}
