% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chad-package.R
\docType{package}
\name{chad-package}
\alias{chad-package}
\title{A lightweight acceptance test framework}
\description{
This provides a way to run acceptance tests for/from R, along with some scripting glue for updating them.
}
\section{Requirements}{

This package needs a \code{diff} program on the search path. Optionally, it can also use \code{git} and \code{colordiff}

Dedicated to Chad Whipkey.
}
\author{
Neal Fultz
}
\references{
\url{http://en.wikipedia.org/wiki/Acceptance_testing}
}
\seealso{
chad
}

