% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{CHAD}
\alias{CHAD}
\title{Constructor method for the CHAngepoint Detector (CHAD) class}
\usage{
CHAD(
  p,
  method = c("mean", "covariance"),
  leading_constant,
  baseline_sd = 1,
  baseline_operatornorm = NA,
  estimate_mean = TRUE,
  constant_penalty = FALSE,
  min_prechange_obs = 1,
  false_alarm_prob = 0.05,
  MC_reps = 100,
  N = 1000,
  pure_R = FALSE
)
}
\arguments{
\item{p}{Dimension of the data}

\item{method}{Test statistic to be used, either \code{"mean"} or
\code{"covariance"}. See details below}

\item{leading_constant}{A numeric vector specifying the leading constants for
the penalty function, or the string "MC" for Monte Carlo
simulation of these
If "MC" is given as input, the arguments \code{false_alarm_prob},
\code{N} and \code{MC_reps} are used in the MC simulation}

\item{baseline_sd}{If \code{method = "mean"}, a numeric vector specifying
the baseline standard deviation for each coordinate of the data
Only applicable when \code{method = "mean"}}

\item{baseline_operatornorm}{If \code{method = "covariance"}, a numeric value
specifying the baseline operator norm of the pre-change covariance matrix.
If \code{baseline_operatornorm = NA}, the operator norm is estimated
in real-time}

\item{estimate_mean}{Boolean indicating whether the data should be estimated
in real time, or taken to be zero}

\item{constant_penalty}{Boolean indicating whether critical values (and
tresholds) should be constant with respect to the sample size}

\item{min_prechange_obs}{Minimum number of of pre-change observations
required for a candidate changepoint location to be considered.
Note that if \code{method = "covariance"}, and
\code{baseline_operatornorm = NA}, it is reccomended to set
\code{min_prechange_obs = 10} to avoid highly variable estimates of the
baseline noise level.}

\item{false_alarm_prob}{False alarm probability for MC simulations
(between 0 and 1). Only required when \code{leading_constant='MC'}}

\item{MC_reps}{Number of Monte Carlo repetitions to estimate the
leading constant(s) of the penalty function(s)
Only required when \code{leading_constant='MC'}}

\item{N}{Maximum sample size of which Monte Carlo simulations are performed.
Only required when \code{leading_constant='MC'}}

\item{pure_R}{Boolean indicating whether the test statistics should be
computed in pure (interpreted) R code instead of in C. Only applicable when
\code{method = "mean"} (the covariance detector contains no compiled code).
The two implementations compute the same quantities and agree up to
floating point error; the option exists to enable run time comparisons
with methods implemented in interpreted languages}
}
\value{
An object of class 'CHAD' with the following attributes:
\itemize{
\item \code{class} - subclass and class
\item \code{p} - Data dimension
\item \code{method} - Method used for online changepoint detection
\item \code{leading_constant} - A vector of leading constant(s) for the
penalty function(s)
\item \code{estimate_mean} - Boolean indicating whether mean estimation
should be performed
\item \code{n_obs} - Number of currently monitored observations
\item \code{n_obs_tot} - Number of observations processed in total
\item \code{baseline_sd} - Vector of standard deviations (for mean changes)
\item \code{baseline_operatornorm} - Baseline operator norm of the covariance
matrix (for covariance changes)
\item \code{statistics} - A vector of test statistics
\item \code{overall_max_statistics} - Vector of largest statistics
recorded so far
\item \code{allstats} - A list of various auxiliary statistics
\item \code{constant_penalty} - Boolean indicating whether penalties and
thresholds are constant with respect to the sample size
\item \code{min_prechange_obs} - Minimum number of pre-change observations
\item \code{status} - Either monitoring or changepoint detected
}
}
\description{
Constructor method for the CHAngepoint Detector (CHAD) class
}
\details{
When \code{method = 'mean'}, the test statistic of
Liu et al. (2021) is used to test for a changepoint. If
\code{method = 'covariance'}, the test statistic of Moen (2024) is used.
}
\examples{
detector_mean <- CHAD::CHAD(
  p = 100, method = "mean",
  leading_constant = c(2, 2)
)
detector_covariance <- CHAD(
  p = 100, method = "covariance",
  leading_constant = c(2)
)
}
\references{
\itemize{
\item Liu et al. (2021) Minimax rates in sparse, high-dimensional change
 point detection. Ann. Statist. 49(2): 1081-1112.
 \item Moen (2024) Minimax rates in variance and covariance changepoint
 testing. arXiv:2405.07757.
}
}
