% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadCheck.R
\name{chadCheck}
\alias{chadCheck}
\title{Run chad scripts for a package}
\usage{
chadCheck()
}
\value{
data frame containing the results
}
\description{
A helper function to run all tests in the current directory, and print out a summary of the results.
}
\seealso{
chadRecur
}

