% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadPlot.R
\name{chadPlot}
\alias{chadPlot}
\title{Print a data url'd image to console}
\usage{
chadPlot()
}
\description{
This functions switches off the current graphics device, then prints out an image 
tag with a data url of the file. This is intended for use in chad notebooks.
}

