% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chadRecur.R
\name{chadRecur}
\alias{chadRecur}
\title{Run Chad scripts on entire directory}
\usage{
chadRecur(path = ".", recursive = TRUE, ask = FALSE, quiet = FALSE)
}
\arguments{
\item{path}{Path of directory to process}

\item{recursive}{include subfolders}

\item{ask}{At each failure, ask if to accept}

\item{quiet}{Suppress logging}
}
\value{
TRUE if all tests passes. Additional test data is also provided as attributes.
}
\description{
This will run \code{chad} on every chadscript in a folder (optionally including subfolders);
}
\seealso{
chad
}

