% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{checkChange}
\alias{checkChange}
\title{Check if a changepoint has occurred}
\usage{
checkChange(detector)
}
\arguments{
\item{detector}{Object of class 'CHAD'}
}
\value{
Updated object \code{detector}
}
\description{
Check if a changepoint has occurred
}
\details{
The function \code{\link{normalizedStatistics}} is applied to check
if a changepoint should be declared.
}
