% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departments}
\alias{departments}
\title{Departments of Chad}
\format{
An \emph{sf} object with 70 features and 16 fields
}
\source{
https://data.humdata.org/dataset/chad-admin-boundaries-admin-1-and-admin-2
}
\usage{
departments
}
\description{
Departments of Chad
}
\examples{
departments

}
\keyword{datasets}
