% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diff.files.R
\name{diff.files}
\alias{diff.files}
\title{Diff two files}
\usage{
\method{diff}{files}(file1, file2, verbose)
}
\arguments{
\item{file1}{a file path}

\item{file2}{another file path}

\item{verbose}{print diff output}
}
\description{
Invoke the \code{diff} UNIX utility to compare two files.
}

