% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffFiles.R
\name{diffFiles}
\alias{diffFiles}
\title{Diff two files}
\usage{
diffFiles(file1, file2, verbose)
}
\arguments{
\item{file1}{a file path}

\item{file2}{another file path}

\item{verbose}{print diff output}
}
\description{
Invoke the \code{diff} UNIX utility to compare two files.
}

