% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{districts}
\alias{districts}
\title{Districts of Chad}
\format{
An \emph{sf} object with 126 features and 4 fields
}
\source{
MSF from OCHA
}
\usage{
districts
}
\description{
Districts of Chad
}
\examples{
districts

}
\keyword{datasets}
