% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/equipChad.R
\name{equipChad}
\alias{equipChad}
\title{Equip a package with Chad}
\usage{
equipChad()
}
\description{
Adds a chad test hook to a packages \code{tests/} and \code{DESCRIPTION}.
Run this from a package's source root directory.
}

