% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{get_grid}
\alias{get_grid}
\title{Function generating the dynamic geometric grid}
\usage{
get_grid(t)
}
\arguments{
\item{t}{Number of observations}
}
\value{
A numeric vector containing the generated grid values
}
\description{
Function generating the dynamic geometric grid
}
\examples{
get_grid(10)
}
