% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{getData}
\alias{getData}
\alias{getData.meanDetector}
\alias{getData.covarianceDetector}
\title{Function for processing a new data point}
\usage{
getData(detector, y_new)

\method{getData}{meanDetector}(detector, y_new)

\method{getData}{covarianceDetector}(detector, y_new)
}
\arguments{
\item{detector}{Object of class 'CHAD'}

\item{y_new}{A new data point. It must be of the same dimension as
specified in the \code{p} attribute of \code{detector}.}
}
\value{
Updated object \code{detector}
}
\description{
Function for processing a new data point
}
\section{Methods (by class)}{
\itemize{
\item \code{getData(meanDetector)}: Process a new data point for subclass 'meanDetector'

\item \code{getData(covarianceDetector)}: Process a new data point for subclass
"covarianceDetector"

}}
