% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{MC_mean}
\alias{MC_mean}
\title{MC simulation for meanDetector}
\usage{
MC_mean(
  p,
  false_alarm_prob,
  constant_penalty,
  MC_reps,
  N,
  estimate_mean = TRUE,
  seed = 123
)
}
\arguments{
\item{p}{data dimension}

\item{false_alarm_prob}{desired false alarm probability}

\item{constant_penalty}{If \code{true}, the thresholding values and penalty
function do not grow with the sample size.}

\item{MC_reps}{number of MC iterations}

\item{N}{maximum number of observed data points}

\item{estimate_mean}{Boolean indicating whether the pre-change mean should
be estimated, or assumed to be zero}

\item{seed}{seed for the RNG}
}
\value{
Vector of length 2 with leading constants for the penalty function
for the meanDetector
}
\description{
Chooses the leading constants for the penalty function for
the 'meanDetector' subclass of 'CHAD'.
}
