% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{mean_penalty_and_threshold}
\alias{mean_penalty_and_threshold}
\title{Function generating penalty and threshold values for
the mean change detector}
\usage{
mean_penalty_and_threshold(t, p)
}
\arguments{
\item{t}{An integer representing the sample size}

\item{p}{An integer representing the data dimension}
}
\value{
A list containing the following elements:
\describe{
  \item{as}{A numeric vector of threshold values}
  \item{nu_as}{Mean centering terms}
  \item{r_values}{A numeric vector of penalty values}
  \item{ss}{A numeric vector of sparsity values}
}
}
\description{
Function generating penalty and threshold values for
the mean change detector
}
\examples{
t <- 100
p <- 10
result <- mean_penalty_and_threshold(t, p)
print(result)

}
