% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_trace.R
\name{nest_data}
\alias{nest_data}
\title{Function to recursively nest data}
\usage{
nest_data(data, parent_id = 0, level_name = NULL)
}
\arguments{
\item{data}{The data to be recursively nested}

\item{parent_id}{The identity of the parent call}

\item{level_name}{The name of the current call}
}
\description{
Function to recursively nest data
}
