% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{normalizedStatistics}
\alias{normalizedStatistics}
\title{Compute maximum ratio between the test statistic(s) divided by the
penalty function(s)}
\usage{
normalizedStatistics(detector)
}
\arguments{
\item{detector}{Object of class 'CHAD'}
}
\value{
maximum of the ratio between the current test statistics and their
respective thresholds.
}
\description{
Compute maximum ratio between the test statistic(s) divided by the
penalty function(s)
}
