% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prefectures}
\alias{prefectures}
\title{Prefectures of LAC region of Chad}
\format{
An \emph{sf} object with 10 features and 18 fields
}
\source{
https://data.humdata.org/dataset/chad-admin-boundaries-admin-1-and-admin-2
}
\usage{
prefectures
}
\description{
Prefectures of LAC region of Chad
}
\examples{
prefectures

}
\keyword{datasets}
