% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{print.CHAD}
\alias{print.CHAD}
\title{Printing methods for the 'CHAD' class}
\usage{
\method{print}{CHAD}(x, ...)
}
\arguments{
\item{x}{object of the 'CHAD' class}

\item{...}{other arguments used in \code{print}}
}
\description{
Printing methods for the 'CHAD' class
}
