% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regions of Chad}
\format{
An \emph{sf} object with 23 features and 13 fields.
}
\source{
https://data.humdata.org/dataset/chad-admin-boundaries-admin-1-and-admin-2
}
\usage{
regions
}
\description{
Regions of Chad
}
\examples{
regions

}
\keyword{datasets}
