% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{reset}
\alias{reset}
\alias{reset.meanDetector}
\alias{reset.covarianceDetector}
\title{Reset an object of the class 'CHAD'}
\usage{
reset(detector)

reset(detector)

\method{reset}{meanDetector}(detector)

\method{reset}{covarianceDetector}(detector)
}
\arguments{
\item{detector}{Object of class 'CHAD'}
}
\value{
Updated object \code{detector}

Updated object \code{detector}
}
\description{
Reset an object of the class 'CHAD'

Reset changepoint detector
}
\section{Methods (by class)}{
\itemize{
\item \code{reset(meanDetector)}: Reset object of subclass 'meanDetector'

\item \code{reset(covarianceDetector)}: Reset object of subclass 'covarianceDetector'

}}
