% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygenize.R
\name{roxygenize}
\alias{roxygenize}
\title{'roxygenize' function is used to automatically construct roxygen headers for a given function.}
\usage{
roxygenize(
  func,
  required_tags = c("author", "param", "return", "note", "examples", "export"),
  provided_tags = list(author = "Nathan C. Layman"),
  model = "gpt-4",
  prompt =
    "You parse a function definition to automatically construct roxygen headers returning only the header."
)
}
\arguments{
\item{func}{Function for which the roxygen documentation needs to be generated.}

\item{required_tags}{Vector of tag names required for the documentation. Default is c("author", "param", "return", "note", "example", "export").}

\item{provided_tags}{List of provided tags in the format (tag_name = "tag_content"). Default is list(author = "Nathan C. Layman").}

\item{model}{The model that will be used for generation. Default is "gpt-4".}

\item{prompt}{The initial prompt that guides the documentation generation. Default is "You parse a function definition to automatically construct roxygen headers."}
}
\value{
Prints out the constructed roxygen header.
}
\description{
'roxygenize' function is used to automatically construct roxygen headers for a given function.
}
\note{
Requires OpenAI API key. Warning will be issued and function will terminate if it is not set.
}
\examples{
\dontrun{
  roxygenize(my_func, required_tags = c("author", "param", "return"), provided_tags = list(author = "Nathan C. Layman"), model = "gpt-4", prompt = "Automatically construct roxygen headers for my_func.")
}

}
\author{
Nathan C. Layman
}
