% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_trace.R
\name{serialize_trace}
\alias{serialize_trace}
\title{Serialze an rlang error trace as a nested JSON object}
\usage{
serialize_trace(e, drop = T)
}
\arguments{
\item{e}{The rlang enhanced error}

\item{drop}{When printing a trace rlang usually drops some rows. Should those be dropped here as well?}
}
\description{
Serialze an rlang error trace as a nested JSON object
}
