% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{settlements}
\alias{settlements}
\title{Settlements of Chad}
\format{
A \emph{tibble} object with 16963 features and 15 fields
}
\source{
https://data.humdata.org/dataset/chad-populated-places-settlements
}
\usage{
settlements
}
\description{
Settlements of Chad
}
\examples{
settlements

}
\keyword{datasets}
