% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/use_chad.R
\name{use_chad}
\alias{use_chad}
\title{Equip a package with Chad}
\usage{
use_chad(pkg = ".")
}
\description{
Adds a chad test hook to a packages \code{tests/} and \code{DESCRIPTION}.
Run this from a package's source root directory.
}

