% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useChad.R
\name{useChad}
\alias{useChad}
\title{Equip a package with Chad}
\usage{
useChad(pkg = ".")
}
\arguments{
\item{pkg}{package path}
}
\description{
Adds a chad test hook to a packages \code{tests/}, and \code{DESCRIPTION} (if you have devtools).
Run this from a package's source root directory.
}

