% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villages}
\alias{villages}
\title{Villages of Chad}
\format{
An \emph{sf} object with 16971 features and 11 fields
}
\source{
TBC
}
\usage{
villages
}
\description{
Villages of Chad
}
\examples{
villages

}
\keyword{datasets}
