% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chad.R
\name{chad}
\alias{chad}
\title{Chad a file}
\usage{
chad(file.name, ask = interactive(), quiet.log = !ask, msg = TRUE,
  echo = FALSE)
}
\arguments{
\item{file.name}{path to a chad script}

\item{ask}{if there is a failure, prompt the user to accept as golden}

\item{quiet.log}{suppress logging messages}

\item{msg}{also log std. err.}

\item{echo}{echo chadscript to screen as it's run -- note: messages cannot be echo'ed.}
}
\description{
Chad a file
}
\section{Options}{

\code{chad.diff} - the diff program to use, defaulting to \code{diff -u}
}

