% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chad.R
\name{chad}
\alias{chad}
\title{Chad a file}
\usage{
chad(file.name, ask = interactive(), quiet = !ask)
}
\arguments{
\item{file.name}{path to a chad script}

\item{ask}{if there is a failure, prompt the user to accept as golden}

\item{quiet}{suppress logging messages}
}
\description{
Chad a file
}
\section{Options}{

\code{chad.diff} - the diff program to use, defaulting to \code{diff -u}
}

