% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chad.r
\name{chad}
\alias{chad}
\title{Chad a file}
\usage{
chad(file.name, ask = interactive())
}
\arguments{
\item{file.name}{path to a chad script}

\item{ask}{if there is a failure, prompt the user to accept as golden}
}
\description{
Chad a file
}
\section{Options}{

\code{chad.color} - if set to TRUE, pipe the diff through \code{colorout}
}

