% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{CHAD}
\alias{CHAD}
\title{Constructor method for the CHAngepoint Detector (CHAD) class}
\usage{
CHAD(
  p,
  method = c("mean", "covariance"),
  leading_constant,
  baseline_sd = 1,
  estimate_mean = TRUE,
  constant_penalty = FALSE,
  false_alarm_prob = 0.05,
  MC_reps = 100,
  N = 1000
)
}
\arguments{
\item{p}{Dimension of the data}

\item{method}{Test statistic to be used, either \code{mean} or
\code{covariance}. See details below}

\item{leading_constant}{A numeric vector specifying the leading constants for
the penalty functions, or the character string 'MC' for Monte Carlo
simulation of these.
If 'MC' is given as input, the \code{false_alarm_prob},
\code{N} and \code{MC_reps} need to be specified}

\item{baseline_sd}{If \code{method = "mean"}, a numeric vector specifying the baseline standard error
for each coordinate of the data. Only applicable for changes in mean}

\item{estimate_mean}{Boolean indicating whether mean-centering should be
performed.}

\item{constant_penalty}{Boolean indicating whether leading constants and
critical values should be constant with respect to the sample size or not}

\item{false_alarm_prob}{False alarm probability for MC simulations (between 0 and 1).
Only required when \code{leading_constant='MC'}}

\item{MC_reps}{Number of Monte Carlo repetitions to estimate the
leading constant(s) of the penalty function(s)
Only required when \code{leading_constant='MC'}}

\item{N}{Maximum sample size of which Monte Carlo simulations are performed.
Only required when \code{leading_constant='MC'}}
}
\value{
An object of class 'CHAD' with the following attributes:
\itemize{
\item \code{class} - Class and subclass
\item \code{p} - Data dimension
\item \code{method} - Method used for online changepoint detection
\item \code{leading_constant} - A vector of leading constant(s) for the
penalty function(s)
\item \code{n_obs} - Number of currently monitored observations
\item \code{n_obs_tot} - Number of observations processed in total
\item \code{baseline_sd} - Vector of standard deviations
\item \code{statistics} - A vector of test statistics
\item \code{overall_max_statistics} - Vector of largest test statistics
recorded
\item \code{allstats} - A list of various auxiliary statistics
\item \code{constant_penalty} - Boolean indicating whether penalties and
thresholds are constant with respect to the sample size
\item \code{status} - Either monitoring or changepoint detected
}
}
\description{
Constructor method for the CHAngepoint Detector (CHAD) class
}
\details{
When \code{method = 'mean'}, the test statistic of
Liu et al. (2021) is used to test for a changepoint. If
\code{method = 'covariance'}, the test statistic of Moen (2024) is used.
}
\examples{
detector_mean <- CHAD::CHAD(p = 100, method = "mean", leading_constant = c(2, 2))
detector_covariance <- CHAD(
  p = 100, method = "covariance",
  leading_constant = c(2)
)
}
\references{
\itemize{
\item Liu et al. (2021) Minimax rates in sparse, high-dimensional change
 point detection. Ann. Statist. 49(2): 1081-1112.
 \item Moen (2024) Minimax rates in variance and covariance changepoint
 testing. arXiv:2405.07757.
}
}
