% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chad.R
\docType{package}
\name{chad}
\alias{chad}
\alias{chad-package}
\title{Datasets for Use in Designing Surveys in Chad}
\description{
Designing surveys require relevant datasets to be used as basis for sample
size calculations, sampling design, survey planning/logistics and survey
implementation. These include datasets on population, lists of sampling
clusters, map datasets for spatial sampling, and previous survey datasets
that can be used for estimating indicator variance and design effects. This
package contains relevant datasets for use in designing surveys in
Chad.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/spatialworks/chad}
  \item \url{https://spatialworks.io/chad/}
  \item Report bugs at \url{https://github.com/spatialworks/chad/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
