% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chadRecur.R
\name{chadRecur}
\alias{chadRecur}
\title{Run Chad scripts on entire directory tree}
\usage{
chadRecur(path = ".", meta = list(start = Sys.time()), ask = FALSE)
}
\arguments{
\item{path}{Path of directory to process}

\item{meta}{metadata to attach to the result}

\item{ask}{At each failure, ask if to accept}
}
\description{
This will run \code{chad} on every chadscript in a folder (including subfolders);
}
\seealso{
chad
}

