% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_last_error.R
\name{explain_last_error}
\alias{explain_last_error}
\title{Use ChatGPT to explain the last error}
\usage{
explain_last_error(
  model = "gpt-3.5-turbo",
  prompt =
    "Give an explanation for the following error in the R programing language and use the provided stack trace to aid the user in resolving the problem. Do not ask for code snippets or any further information.",
  show_trace = F,
  show_trace_JSON = F
)
}
\arguments{
\item{model}{Which version of chatgpt would should the error be submitted to?}

\item{prompt}{What system prompt should be used to interpret the error and stack trace?}

\item{show_trace}{Whether to print the rlang trace}

\item{show_trace_JSON}{Whether to display the serialized trace}
}
\description{
Use ChatGPT to explain the last error
}
