% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{new_covariance}
\alias{new_covariance}
\title{Constructor of subclass 'covarianceDetector' of class 'CHAD'}
\usage{
new_covariance(p, leading_constant, estimate_mean, constant_penalty)
}
\arguments{
\item{p}{Data dimension}

\item{leading_constant}{A single numeric value}

\item{estimate_mean}{Boolean indicating whether mean-centering should be
performed.}

\item{constant_penalty}{Boolean indicating whether leading constants and
critical values should be constant with respect to the sample size or not}
}
\value{
An object of subclass 'covarianceDetector' of class 'CHAD'
}
\description{
Constructor of subclass 'covarianceDetector' of class 'CHAD'
}
\details{
Implements the test statistic of Moen (2024) for dense
changes in covariance
}
\examples{
detector <- CHAD::new_covariance(
  p = 100, leading_constant = c(2),
  estimate_mean = TRUE, constant_penalty = FALSE
)
}
\references{
\itemize{
 \item Moen (2024) Minimax rates in variance and covariance changepoint
 testing. arXiv:2405.07757.
}
}
