% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAD.R
\name{new_covariance}
\alias{new_covariance}
\title{Constructor of subclass 'covarianceDetector' of class 'CHAD'}
\usage{
new_covariance(p, leading_constant, estimate_mean, constant_penalty)
}
\arguments{
\item{p}{Data dimension}

\item{leading_constant}{A numeric vector of length 1}
}
\value{
An object of subclass 'covarianceDetector' of class 'CHAD'
}
\description{
Constructor of subclass 'covarianceDetector' of class 'CHAD'
}
\details{
Implements the test statistic of Moen (2024) for dense
changes in covariance online
}
\examples{
detector <- new_covariance(p=100, leading_const = c(2),
            estimate_mean = TRUE, constant_penalty = FALSE)
}
\references{
\itemize{
 \item Moen (2024) Minimax rates in variance and covariance changepoint
 testing. arXiv:2405.07757.
}
}
