% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize_trace.R
\name{serialize_trace}
\alias{serialize_trace}
\title{Serialize Error Trace to JSON}
\usage{
serialize_trace(e, drop = T)
}
\arguments{
\item{e}{An error object generated by `rlang`. It must contain a `trace` component, which is a dataframe-like structure.}

\item{drop}{A boolean value indicating whether to filter and include only visible trace entries (i.e., where `visible == TRUE`).
Default is TRUE.}
}
\value{
A nested JSON string representing the structured error trace.
}
\description{
This function serializes the trace from an `rlang` error object into a nested JSON structure.
The trace is filtered, processed, and converted into a nested format using the `nest_data` function,
which organizes the function calls based on their parent-child relationships. The resulting trace is then
serialized into a JSON object for easy inspection or logging.
}
\note{
This function requires the `dplyr` and `jsonlite` packages for filtering and JSON serialization. It relies on
the `nest_data` function to handle nesting of trace data before serialization.
}
\examples{
tryCatch(
    sqrt("a"),
    error = function(e) {
    print(serialize_trace(e))
})

}
\author{
Nathan C. Layman
}
