% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{settlements}
\alias{settlements}
\title{Settlements of Chad}
\format{
A \emph{tibble} object with 11524 rows and 19 columns
}
\source{

}
\usage{
settlements
}
\description{
Settlements of Chad
}
\examples{
settlements

}
\keyword{datasets}
